+++ tms-api/src/main/java/pu/master/tmsapi/models/requests/UserRequest.java
@@ -28,9 +28,9 @@ public class UserRequest
 
     private LocalDate dateLastModifiedAt;
 
-    private List<Integer> projects;
+    private List<Long> projects;
 
-    private Set<Integer> roles;
+    private Set<Long> roles;
 
 
     public String getUsername()
@@ -150,26 +150,26 @@ public class UserRequest
     }
 
 
-    public List<Integer> getProjects()
+    public List<Long> getProjects()
     {
         return projects;
     }
 
 
-    public UserRequest setProjects(final List<Integer> projects)
+    public UserRequest setProjects(final List<Long> projects)
     {
         this.projects = projects;
         return this;
     }
 
 
-    public Set<Integer> getRoles()
+    public Set<Long> getRoles()
     {
         return roles;
     }
 
 
-    public UserRequest setRoles(final Set<Integer> roles)
+    public UserRequest setRoles(final Set<Long> roles)
     {
         this.roles = roles;
         return this;
